+++ srcs/percent_p_and_o.c
@@ -6,7 +6,7 @@
 /*   By: kemethen <kemethen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/16 12:36:31 by kemethen          #+#    #+#             */
-/*   Updated: 2019/04/15 18:57:20 by kemethen         ###   ########.fr       */
+/*   Updated: 2019/04/18 13:48:24 by kemethen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int		msize_hexa_ptr(unsigned long n)
 	return (i);
 }
 
-size_t	mbase_eight(unsigned long n)
+size_t	mbase_eight(unsigned int n)
 {
 	size_t	j;
